로그인전보드수정

+++ WebContent/board/board_content.jsp
@@ -71,7 +71,11 @@
 							<%-- 조회수 --%>
 							<span>조회수: ${dto.getBoard_hit() }</span> <span>추천수: <img
 								src="img/thumbup.png" width="30" height="30" id="thumbs"
-								onclick="thumbsClick()"> <span class="thumbs"></span>
+								onclick="if(${loginCheck } == 0){
+									alert('로그인이 필요합니다');
+							}else {
+								thumbsClick();
+							}"><span class="thumbs"></span>
 							</span>
 
 						</div>
@@ -86,20 +90,30 @@
 				<span>삭제된 게시물입니다</span>
 			</c:if>
 			<br> <input type="button" value="글 수정"
-				onclick="if(${loginCheck } == 0){
-						alert('로그인이 필요합니다');
-						location.href='member/login.jsp'
-					}else {
-						location.href='board_modify.do?no=${dto.getBoard_index() }&page=${Page }&type=${dto.getBoard_type() }'
-					}">
-			<input type="button" value="글 삭제"
-				onclick="if(${loginCheck } == 0){
-						alert('로그인이 필요합니다');
-						location.href='member/login.jsp'
-					}else {
-						if(confirm('정말로 삭제하시겠습니까?')) {
-							location.href='board_delete.do?no=${dto.getBoard_index() }&page=${Page }&type=${dto.getBoard_type() }'
-							}else { retrun;}}">
+				    onclick="if(${loginCheck } == 0){
+				            alert('로그인이 필요합니다');
+				            location.href='<%=request.getContextPath() %>/member/login.jsp'
+				        }else {
+				            if('${dto.getBoard_writer_id()}' == '${member_id}') {
+				                location.href='board_modify.do?no=${dto.getBoard_index() }&page=${Page }&type=${dto.getBoard_type() }'
+				            }else {
+				                alert('본인이 쓴 글이 아닙니다!');
+				            }
+				        }">
+				        
+				<input type="button" value="글 삭제"
+				    onclick="if(${loginCheck } == 0){
+				            alert('로그인이 필요합니다');
+				            location.href='<%=request.getContextPath() %>/member/login.jsp'
+				        }else {
+				            if('${dto.getBoard_writer_id()}' == '${member_id}') {
+				                if(confirm('정말로 삭제하시겠습니까?')) {
+				                location.href='board_delete.do?no=${dto.getBoard_index() }&page=${Page }&type=${dto.getBoard_type() }'
+				                }else { return;}
+				            }else {
+				                alert('본인이 쓴 글이 아닙니다!');
+				            }
+				        }">
 			<input type="button" value="전체목록"
 				onclick="location.href='board_list.do?type=${dto.getBoard_type() }'">
 			<br> <br>
